pyproject.toml
[tool.poetry]
name = "cardyard-bot"
version = "0.1.0"
description = ""
authors = ["zaidejaz <zaidejaz521@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
colorlog = "^6.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
